BookManagement_SERVLETS: Move all book's servlets to package "servlet.book"
