+++ src/servlets/book/AddServlet.java
@@ -1,18 +1,17 @@
-package servlets;
+package servlets.book;
 
 import java.io.IOException;
-
 import java.io.PrintWriter;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import beans.Book;
 import daos.BookDAO;
-@WebServlet("/EditServlet")
-public class EditServlet extends HttpServlet {
+@WebServlet("/AddServlet")
+public class AddServlet extends HttpServlet {
 
 	private static final long _SERIAL_VERSION_UID = 1L;
 
@@ -30,32 +29,27 @@ public class EditServlet extends HttpServlet {
 		processRequest(request, response);
 
 		PrintWriter out = response.getWriter();
-		String id = request.getParameter("id");
+
 		String name = request.getParameter("name");
 		String author = request.getParameter("author");
-		Book book = new Book();
-		book.setId(id);
-		book.setName(name);
-		book.setAuthor(author);
 		try {
 
-			int status = BookDAO.update(book);
-
-			if(status > 0) {
+			boolean status = BookDAO.save(name, author);
+			if(status) {
 
 				response.sendRedirect("view");
 			}
 			else {
 
-				out.println("Sorry! unable to update this book");
+				out.println("Sorry! unable to add this book");
 			}
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 
-			out.println("Caught error when updating: " + e);
+			out.println("Caught error when adding:" + e);
+
 		}
+
 		out.close();
 	}
 
 }
-
